[#BUGFIX] getTypoScriptDataHook did not work as expected in all cases
[+FEATURE] Added T3X_mm_forum_import_t3blog.t3x: Very simple temporarily extension for importing ALL existing t3blog comments on buttonclick. Be aware that this solution is very simple and should be tested in development environments.
For importing process select a fe_user in mm_forum install backend module for Blog comments, that you will delete after the import. This user id will be used for the post creation process. It is recommended to delete the user after the import process because the post itself is displayed as a guest account but the profile of the fe_user will show all comments that does not really belong to the fe_user.

git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/mm_forum_blog/trunk@33865 735d13b6-9817-0410-8766-e36946ffe9aa